% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{agro_to_odk_choices}
\alias{agro_to_odk_choices}
\title{Convert from AgroFIMS format to ODK choices description}
\usage{
agro_to_odk_choices(traitlist)
}
\arguments{
\item{traitlist}{data.frame table of al the list of traits}

\item{dictionary}{data.frame dictionary of all the equivalences between AgroFIMS and mobile apps attributes}
}
\description{
convert AgroFIMS trait list data to ODK choices-sheet format
}
