% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{agro_to_odk_survey}
\alias{agro_to_odk_survey}
\title{Convert from AgroFIMS format to ODK survey - trait list format}
\usage{
agro_to_odk_survey(traitlist, dictionary)
}
\arguments{
\item{traitlist}{data.frame table of al the list of traits}

\item{dictionary}{data.frame dictionary of all the equivalences between AgroFIMS and mobile apps attributes}
}
\description{
Convert from AgroFIMS format to ODK survey - trait list format
}
