% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{assign_expunit}
\alias{assign_expunit}
\title{Assign  experimental unit}
\usage{
assign_expunit(
  .info_data,
  dbattr_mea = "length_p",
  dbattr_expunit = "length_p_unit"
)
}
\arguments{
\item{.info_data}{}

\item{expmea}{}

\item{expunit}{}
}
\description{
Assign  experimental unit
}
