% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_design_factor.R
\name{calc_fert_design}
\alias{calc_fert_design}
\title{Calculation of nutrient and fertilizer amount in experimental design factors}
\usage{
calc_fert_design(fertilizer, factorId = NULL)
}
\arguments{
\item{fertilizer}{data.frame fertilizer table}

\item{factorId}{integer factor database id , internally coded by \code{groupId} in AGROFIMS Database.}
}
\description{
Calculation of nutrient and fertilizer amount in experimental design factors
}
\examples{
\dontrun{
library(ragrofims)
library(ragapi)
fertilizer <-get_agrofims_designprod(25)
out <- calc_fert_design(fertilizer,2)
}
}
