% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_design_factor.R
\name{calc_nut_design}
\alias{calc_nut_design}
\title{Calculation of nutrient amount in experimental design factors}
\usage{
calc_nut_design(fertilizer, factorId = NULL)
}
\arguments{
\item{fertilizer}{data.frame fertilizer table}

\item{factorId}{integer factor database id , internally coded by \code{groupId} in AGROFIMS Database.}
}
\description{
Calculation of nutrient amount in experimental design factors
}
\examples{
\dontrun{
library(ragrofims)
library(ragapi)
 factors_data <- ragapi::ag_get_edsfactors_expsiteId(expsiteDbId = expsiteId,
                   format = "data.frame",
                   serverURL = serverURL, version = version
                   )
fertilizer <-get_agrofims_designprod(25)
out <- calc_fert_design(fertilizer, factors_data, 4)
}
}
