% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_measurements.R
\name{calc_nutamount}
\alias{calc_nutamount}
\title{Calculation of nutrient amounts based on fertilizer products ()}
\usage{
calc_nutamount(fertilizer)
}
\arguments{
\item{fertilizer}{data.frame fertilizer table which includes fertilizer products and nutrient elements}
}
\description{
Calculation of nutrient amounts based on fertilizer products ()
}
