% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_measurements.R
\name{calc_prodamount}
\alias{calc_prodamount}
\title{Calculation of product amounts based on nutrient concentration}
\usage{
calc_prodamount(nutrient)
}
\arguments{
\item{nutrient}{data.frame nutrient table which includes fertilizer products and nutrient split elements}
}
\description{
Calculation of product amounts based on nutrient concentration
}
