% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_measurements.R
\name{calc_prodnut_split}
\alias{calc_prodnut_split}
\title{Calculation of nutrient concentration per split application}
\usage{
calc_prodnut_split(nut_split, prodnut)
}
\arguments{
\item{nut_split}{numeric nutrient element per split}

\item{prodnut}{numeric nutrient element in product}
}
\description{
calculae nutrient concentration in fertilizer products
}
