% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_traitlist.R
\name{check_duplicates_traitlist}
\alias{check_duplicates_traitlist}
\title{Check duplicates on trait list tables}
\usage{
check_duplicates_traitlist(traitlist)
}
\arguments{
\item{traitlist}{data.frame trait list table}
}
\description{
check whether exist duplicates variables or measurements in the trait list table
}
