% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_expdesign_elements.R
\name{ck_factor_names}
\alias{ck_factor_names}
\title{Check factor names}
\usage{
ck_factor_names(.data)
}
\arguments{
\item{.data}{data.frame experimental data which includes factors and levels retrieved from AGROFIMS database via API.}
}
\description{
Check factor names
}
