% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_expdesign_elements.R
\name{ck_level_values}
\alias{ck_level_values}
\title{Check missing values in levels}
\usage{
ck_level_values(.data)
}
\arguments{
\item{.data}{data.frame experimental data which includes factors and levels retrieved from AGROFIMS database via API.}
}
\description{
Check missing values in levels
}
