% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding_metadata.R
\name{clean_fundagency}
\alias{clean_fundagency}
\title{Clean labels in order to export in MS Excel format}
\usage{
clean_fundagency(.data)
}
\arguments{
\item{.data}{data.frame table with}

\item{attribute}{character vector of databases attributes}

\item{meta_dbattributes}{metadata attributes and labels in AgroFIMS}
}
\description{
Functionality to clean and poolish the AGROFIMS API response.

clean api response attributes by AgroFIMS label in order to export the fieldbook file
}
\examples{
\dontrun{
.data <- ag_get_fundagency_studyId(studyDbId = 28,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data <- .replace_other_attribute_funding(.data, "fundAgencyTypeId", "fundAgencyTypeOther")
.data <- clean_fundagency(.data)
}
}
\author{
Omar Benites
}
