% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel_metadata.R
\name{clean_personnel}
\alias{clean_personnel}
\title{Clean labels in order to export in MS Excel format}
\usage{
clean_personnel(.data)
}
\arguments{
\item{.data}{data.frame table with}
}
\description{
Functionality to clean and poolish the AGROFIMS API response.

clean api response attributes by AgroFIMS label in order to export the fieldbook file
}
\examples{
\dontrun{
.data <- ag_get_personnel_studyId(studyDbId = 28,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data <- .replace_other_attribute_personnel(.data, "fundAgencyTypeId", "fundAgencyTypeOther")
.data <- clean_personnel(.data)
}
}
\author{
Omar Benites
}
