% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsite_metadata.R
\name{clean_sitedesc}
\alias{clean_sitedesc}
\title{Clean values in experiment-site information}
\usage{
clean_sitedesc(.data)
}
\arguments{
\item{.data}{data to clean up \code{Other} values}

\item{attribute}{attribute of the database. Attribute name from AgroFIMS database where a user input is stored}

\item{other_attribute}{Other attribute name related to \code{attribute} parameter used to store \code{Other} values or non-standardized inputs.}
}
\description{
Tipically values users fill forms or combos with standarized inputs in AgroFIMS. However, sometimes ontologized inputs are not enough.
For this reason, they select the \code{Other} value in order to fill options that have not been maped in the Agronomic Ontology in AgroFIMS
}
\examples{
\dontrun{
.data <- ag_get_sitedesc_studyId(studyDbId = 21,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data. <- clean_sitedesc(.data, "experimentType", "experimentTypeOther")
}
}
\author{
Omar Benites
}
