% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropsite_metadata.R
\name{convert_to_xlsx_cropdesc}
\alias{convert_to_xlsx_cropdesc}
\title{Convert crop information response to AgroFIMS Excel R-table}
\usage{
convert_to_xlsx_cropdesc(.data, meta_dbattributes)
}
\arguments{
\item{.data}{data.frame table with crop information metadata}

\item{meta_dbattributes}{metadata attributes and labels in AgroFIMS}
}
\description{
Transform an API table to Excel table format

set and modify databases attributes by AgroFIMS label in order to export the fieldbook file
}
\author{
Omar Benites
}
