% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdetails_metadata.R
\name{convert_to_xlsx_expdetails}
\alias{convert_to_xlsx_expdetails}
\title{Convert API-data.frame response to AgroFIMS Excel logic table}
\usage{
convert_to_xlsx_expdetails(.data, meta_dbattributes)
}
\arguments{
\item{.data}{data.frame table with}

\item{meta_dbattributes}{metadata attributes and labels in AgroFIMS}

\item{attribute}{character vector of databases attributes}
}
\description{
Transform an API table to Excel table format

set and modify databases attributes by AgroFIMS label in order to export the fieldbook file
}
\author{
Omar Benites
}
