% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funding_metadata.R
\name{convert_to_xlsx_fundagency}
\alias{convert_to_xlsx_fundagency}
\title{Convert API-data.frame response to AgroFIMS Excel logic table}
\usage{
convert_to_xlsx_fundagency(.data, meta_dbattributes)
}
\arguments{
\item{.data}{data.frame table with}

\item{meta_dbattributes}{metadata attributes and labels in AgroFIMS}

\item{attribute}{character vector of databases attributes}
}
\description{
Transform an API table to Excel table format

set and modify databases attributes by AgroFIMS label in order to export the fieldbook file
}
\examples{
\dontrun{
.data <- ag_get_fundagency_studyId(studyDbId = 28,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data <- .replace_other_attribute_funding(.data, "fundAgencyTypeId", "fundAgencyTypeOther")
.data <- clean_fundagency(.data)
.data <- convert_to_xlsx_fundagency(.data)
}
}
\author{
Omar Benites
}
