% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel_metadata.R
\name{convert_to_xlsx_personnel}
\alias{convert_to_xlsx_personnel}
\title{Convert API-data.frame response to AgroFIMS Excel logic table}
\usage{
convert_to_xlsx_personnel(.data, meta_dbattributes = NULL)
}
\arguments{
\item{.data}{data.frame table with}

\item{meta_dbattributes}{metadata attributes and labels in AgroFIMS}
}
\description{
Transform an API table to Excel table format

set and modify databases attributes by AgroFIMS label in order to export the fieldbook file
}
\examples{
\dontrun{
.data <- ag_get_personnel_studyId(studyDbId = 28,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data <- .replace_other_attribute_personnel(.data, "fundAgencyTypeId", "fundAgencyTypeOther")
.data <- clean_personnel(.data)
.data <- convert_to_xlsx_personnel(.data)
}
}
\author{
Omar Benites
}
