% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsite_metadata.R
\name{convert_to_xlsx_sitedesc}
\alias{convert_to_xlsx_sitedesc}
\title{Convert API-data.frame response to AgroFIMS Excel logic table}
\usage{
convert_to_xlsx_sitedesc(.data, meta_dbattributes)
}
\arguments{
\item{.data}{data.frame table with experimental details metadata}

\item{meta_dbattributes}{metadata attributes and labels in AgroFIMS}
}
\description{
Transform an API table to Excel table format

set and modify databases attributes by AgroFIMS label in order to export the fieldbook file
}
\author{
Omar Benites
}
