% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropmea_vars.R
\name{cr_cmeasurement_fbook}
\alias{cr_cmeasurement_fbook}
\title{Crop measurement fieldbook}
\usage{
cr_cmeasurement_fbook(design, traitlist)
}
\arguments{
\item{design}{design}

\item{traitlist}{traitlist}
}
\description{
Crop measurement fieldbook
}
