% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_utils.R
\name{cr_experimental_design}
\alias{cr_experimental_design}
\title{Create experimental design in AGROFIMS}
\usage{
cr_experimental_design(
  design,
  rep = 2,
  block = 2,
  trt = 2,
  ntrt = NULL,
  fnames = NULL,
  flevels = NULL,
  mplot = NULL,
  splot = NULL,
  ssplot = NULL,
  rowf = NULL,
  colf = NULL
)
}
\arguments{
\item{design}{experimental design abbreviation.}

\item{rep}{replication}

\item{block}{blocks}

\item{trt}{treatments}

\item{ntrt}{number of treatments}

\item{fnames}{list of factors names}

\item{flevels}{list of levels according to previous factors}

\item{mplot}{main plot}

\item{splot}{sub plot}

\item{ssplot}{sub-sub plot}

\item{rowf}{row factor}

\item{colf}{col factor}
}
\description{
create experimental design in agrofims through user interface inputs
}
