% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{cr_fbapp}
\alias{cr_fbapp}
\title{Create Field Book App template}
\usage{
cr_fbapp(design)
}
\arguments{
\item{design}{data.frame statistical design object including crop measurements}
}
\description{
Create Field Book App template
}
