% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manprac_vars.R
\name{cr_mangprac_fbook}
\alias{cr_mangprac_fbook}
\title{Create Management Pratices fieldbook}
\usage{
cr_mangprac_fbook(design, traitlist)
}
\arguments{
\item{design}{data.frame experimental design layout}

\item{traitlist}{data.frame trait list table which includes all management practices variables. Ex. units, evalutation per season/plot.}
}
\description{
create a management practices field book based on trait list tables
}
