% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pheno_vars.R
\name{cr_phenology_fbook}
\alias{cr_phenology_fbook}
\title{Create Phenology measurement fieldbook}
\usage{
cr_phenology_fbook(design, traitlist)
}
\arguments{
\item{design}{design}

\item{traitlist}{traitlist}
}
\description{
create phenology book(s) based on trait list tables
}
