% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_vars.R
\name{cr_soil_fbook}
\alias{cr_soil_fbook}
\title{Create Phenology measurement fieldbook}
\usage{
cr_soil_fbook(design, traitlist)
}
\arguments{
\item{design}{design}

\item{traitlist}{traitlist}
}
\description{
create soil field book(s) based on trait list tables
}
