% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.get_other_rowpos_expdetails}
\alias{.get_other_rowpos_expdetails}
\title{Get other row position in expdetails data}
\usage{
.get_other_rowpos_expdetails(.data, other_attribute)
}
\arguments{
\item{.data}{data to clean up \code{Other} values}

\item{other_attribute}{Other attribute name related to \code{attribute} parameter used to store \code{Other} values or non-standardized inputs.}
}
\description{
Get other row position in expdetails data
}
\examples{
\dontrun{
.data <- ag_get_expdetails_studyId(studyDbId = 21,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data. <- .get_other_position_expdetails(.data, "experimentTypeOther")
}
 
}
\author{
Omar Benites
}
