% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.modify_labels_expdetails}
\alias{.modify_labels_expdetails}
\title{Modify labels for metadata table}
\usage{
.modify_labels_expdetails(.data, meta_dbattributes)
}
\arguments{
\item{.data}{data.frame table with}

\item{meta_dbattributes}{metadata attributes and labels in AgroFIMS}

\item{attribute}{character vector of databases attributes}
}
\description{
set and modify databases attributes by AgroFIMS label in order to export the fieldbook file
}
\author{
Omar Benites
}
