% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.remove_other_exdetails}
\alias{.remove_other_exdetails}
\title{Remove other values in expdtails data}
\usage{
.remove_other_exdetails(.data, other_attribute)
}
\arguments{
\item{.data}{data to clean up \code{Other} values}

\item{other_attribute}{Other attribute name related to \code{attribute} parameter used to store \code{Other} values or non-standardized inputs.}
}
\description{
Remove other values in expdtails data
}
\examples{
\dontrun{
.data <- ag_get_expdetails_studyId(studyDbId = 21,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data. <- .remove_other_exdetails(.data, "experimentTypeOther")
}
}
\author{
Omar Benites
}
