% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.replace_other_attribute_expdetails}
\alias{.replace_other_attribute_expdetails}
\title{Replace attributes that has others values in expdetails}
\usage{
.replace_other_attribute_expdetails(.data, attribute, other_attribute)
}
\arguments{
\item{.data}{data to clean up \code{Other} values}

\item{attribute}{attribute of the database. Attribute name from AgroFIMS database where a user input is stored}

\item{other_attribute}{Other attribute name related to \code{attribute} parameter used to store \code{Other} values or non-standardized inputs.}
}
\description{
Replace attributes that has others values in expdetails
}
\examples{
\dontrun{
.data <- ag_get_expdetails_studyId(studyDbId = 21,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data. <- .replace_other_attribute_expdetails(.data, "experimentType" ,"experimentTypeOther")
}
  
}
\author{
Omar Benites
}
