% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/projlead_metadata.R
\name{.replace_other_attribute_projlead}
\alias{.replace_other_attribute_projlead}
\title{Replace attributes that has others values in project entity}
\usage{
.replace_other_attribute_projlead(.data, attribute, other_attribute)
}
\arguments{
\item{.data}{data to clean up \code{Other} values}

\item{attribute}{attribute of the database. Attribute name from AgroFIMS database where a user input is stored}

\item{other_attribute}{Other attribute name related to \code{attribute} parameter used to store \code{Other} values or non-standardized inputs.}
}
\description{
Tipically, users type values that are not mapped in the agronomy ontology. For this reason, the API response retrieve additional information
that should ensemble in a data structure.
}
\examples{
\dontrun{
.data <- ag_get_projlead_studyId(studyDbId = 28,format = "data.frame",
                                   serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
                                   version ="/0212/r")
.data <- .replace_other_attribute_projlead(.data, "fundAgencyTypeId", "fundAgencyTypeOther")
}
}
\author{
Omar Benites
}
