% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropsite_metadata.R
\name{filter_cropdesc_data}
\alias{filter_cropdesc_data}
\title{Filter crop information according to croppping type.}
\usage{
filter_cropdesc_data(.data)
}
\arguments{
\item{.data}{crop information data}
}
\description{
every croppping type has specific information. For this reason, this function filter this information according to each cropping type.
}
