% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{filter_dsginfo_design}
\alias{filter_dsginfo_design}
\title{Filter experimental design information by design}
\usage{
filter_dsginfo_design(dsginfo, design, .factors_data)
}
\arguments{
\item{dsginfo}{experimetanl design information}

\item{design}{experimental design}

\item{.factors_data}{table of factors retrived by AgroFIMS API.}
}
\description{
According to different
}
