% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{filter_expunitinfo_design}
\alias{filter_expunitinfo_design}
\title{Filter experimental units}
\usage{
filter_expunitinfo_design(dsginfo, expunit, design)
}
\description{
Filter experimental unit values from AGROFIMS experiments
}
