% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_design_factor.R
\name{filter_factors_factorId}
\alias{filter_factors_factorId}
\title{Filter factors by group}
\usage{
filter_factors_factorId(.data, factorId = NULL)
}
\arguments{
\item{.data}{data about factor in experimental designs}

\item{factorId}{character factorId or group number}
}
\description{
Every factor in the AgroFIMS interface has an numeric id that indicates a order or position. This id or position is used to filter and select certain factor that has additional information. 
Example: Fertilizer type and amount, or, nutrient content type and amount.
}
