% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{flatten_manprac_actualplan}
\alias{flatten_manprac_actualplan}
\title{Flatten management practices table to retrieve actual plan evaluations (in trait list)}
\usage{
flatten_manprac_actualplan(traitlist_dt)
}
\arguments{
\item{traitlist_dt}{data.frame trait list table}
}
\description{
Flatten management practices table to retrieve actual plan evaluations (in trait list)
}
