% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_design_factor.R
\name{get_agrofims_designprod}
\alias{get_agrofims_designprod}
\title{Get fertilizer information about factor in experimental designs}
\usage{
get_agrofims_designprod(
  expsiteId = NULL,
  format = "data.frame",
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0291/r"
)
}
\arguments{
\item{expsiteId}{experiment-site Id or expsiteId}

\item{format}{character type of data structure. By default \code{data.frame}.}

\item{serverURL}{database server URL}

\item{version}{api version}
}
\description{
Get fertilizer information about factor in experimental designs
}
\examples{
\dontrun{
library(ragapi)
library(ragrofims)
out<- get_agrofims_designprod(25, "data.frame")
}
}
