% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_measurements.R
\name{get_agrofims_fertproducts}
\alias{get_agrofims_fertproducts}
\title{Get fertilizer product data from AgroFIMS platform}
\usage{
get_agrofims_fertproducts(
  expsiteId = NULL,
  format = "data.frame",
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0291/r"
)
}
\arguments{
\item{expsiteId}{experiment-site Id or expsiteId}

\item{format}{type of data structure format}

\item{serverURL}{database server URL}

\item{version}{api version}
}
\description{
Get fertilizer product data from AgroFIMS platform
}
