% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/traitlist_variables.R
\name{get_agrofims_traitlist}
\alias{get_agrofims_traitlist}
\title{Get metadata from experimental details}
\usage{
get_agrofims_traitlist(
  expsiteId = NULL,
  format = "data.frame",
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0248/r"
)
}
\arguments{
\item{expsiteId}{experiment-site Id or expsiteId}

\item{format}{type of data structure format}

\item{serverURL}{database server URL}

\item{version}{api version}
}
\description{
Get metadata from experimental details
}
