% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropsite_metadata.R
\name{get_cropdesc_metadata}
\alias{get_cropdesc_metadata}
\title{Get metadata of crop information from AGROFIMS experiments}
\usage{
get_cropdesc_metadata(
  expsiteId = NULL,
  format = NULL,
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0291/r",
  meta_dbattributes = NULL
)
}
\arguments{
\item{expsiteId}{experimente site ID}

\item{format}{type of data structure format}

\item{serverURL}{database server URL}

\item{version}{api version}

\item{meta_dbattributes}{data dictionary of metadata. It includes equivalences between excel and database names.}
}
\description{
get metadata from experimental details
}
