% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{get_dsginfo_data}
\alias{get_dsginfo_data}
\title{Get metadata from experimental desing information}
\usage{
get_dsginfo_data(
  expsiteId = NULL,
  format = c("json", "list", "data.frame"),
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0291/r",
  meta_dbattributes = NULL
)
}
\arguments{
\item{expsiteId}{expsiteDbId or experimente site ID}

\item{format}{type of data structure format}

\item{serverURL}{database server URL}

\item{version}{api version}

\item{meta_dbattributes}{data dictionary of metadata. It includes equivalences between excel and database names.}
}
\description{
get metadata from experimental details
}
\examples{
\dontrun{
meta_dbattributes <- gsheet::gsheet2tbl("https://docs.google.com/spreadsheets/d/124fPX0f_J9Ws-f4ZgSX6AyXQVbCZY50nMsuaFfPaDWg/edit#gid=997278569")
get_dsginfo_data(expsiteId = 22, format="data.frame", serverURL = "https://research.cip.cgiar.org/agrofims/api/dev", version= "/0291/r", meta_dbattributes)
}
}
