% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{get_expdesign_abbr}
\alias{get_expdesign_abbr}
\title{Get experimental design abbreviation}
\usage{
get_expdesign_abbr(.data, col_name = "parametercode")
}
\arguments{
\item{.data}{experimental design data}

\item{col_name}{column whic has experimental design abbreviation}
}
\description{
Get experimental design abbreviation
}
