% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{get_expdesign_factors}
\alias{get_expdesign_factors}
\title{Get experimental design factors}
\usage{
get_expdesign_factors(.data)
}
\arguments{
\item{.data}{experimental design data}
}
\description{
Get experimental design factors
}
