% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{get_expdesign_params}
\alias{get_expdesign_params}
\title{Get experimental design parameters}
\usage{
get_expdesign_params(.data, abbr = "frcbd")
}
\arguments{
\item{.data}{experimental design data}

\item{abbr}{column whic has experimental design abbreviation}
}
\description{
Get experimental design parameters
}
