% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdetails_metadata.R
\name{get_expdetails_metadata}
\alias{get_expdetails_metadata}
\title{Get metadata from experimental details}
\usage{
get_expdetails_metadata(
  studyId = NULL,
  format = NULL,
  serverURL = NULL,
  version = NULL,
  meta_dbattributes = NULL
)
}
\arguments{
\item{studyId}{study or experiment ID}

\item{format}{type of data structure format}

\item{serverURL}{database server URL}

\item{version}{api version}

\item{meta_dbattributes}{data dictionary of metadata. It includes equivalences between excel and database names.}
}
\description{
get metadata from experimental details
}
