% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{get_experimental_design}
\alias{get_experimental_design}
\title{Get experimental design from AgroFIMS via Agronomic API}
\usage{
get_experimental_design(
  expsiteId = NULL,
  format = c("json", "list", "data.frame"),
  serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",
  version = "/0291/r"
)
}
\arguments{
\item{expsiteId}{character experiment site ID from AgroFIMS database}

\item{format}{data format: json, list and data.frame}

\item{serverURL}{URL of the agrofims server}

\item{version}{version of the call. By default version \code{0212}.}
}
\description{
Get experimental design from AgroFIMS via Agronomic API
}
\examples{
\dontrun{
#Get experimental design
out <- get_design_data(expsiteDbId = 6, 
                       format=c("json","list","data.frame"),
                       serverURL="https://research.cip.cgiar.org/agrofims/api/dev", 
                       version = "/0233/r"
                       )

}
}
