% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{get_factorial_levels}
\alias{get_factorial_levels}
\title{Get factorial levels from experimental designs}
\usage{
get_factorial_levels(.data)
}
\arguments{
\item{.data}{experimental design data}
}
\description{
Get factorial levels from experimental designs
}
