% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_measurements.R
\name{get_fertproducts_crop}
\alias{get_fertproducts_crop}
\title{Get fertilizer product table by crop}
\usage{
get_fertproducts_crop(fertproducts, crop)
}
\arguments{
\item{fertproducts}{data.frame fertilizer product table}

\item{crop}{character crop name}
}
\description{
Get fertilizer product table by crop
}
