% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{get_nonfactorial_levels}
\alias{get_nonfactorial_levels}
\title{Get levels from non-full-factorial experiments.}
\usage{
get_nonfactorial_levels(.info_data)
}
\arguments{
\item{.info_data}{experimetal design data}
}
\description{
Extract levels from non-full factorial experiments such as RCBD and CRD.
}
