% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/personnel_metadata.R
\name{get_personnel_metadata}
\alias{get_personnel_metadata}
\title{Get metadata from personnel metadata agency}
\usage{
get_personnel_metadata(
  studyId = NULL,
  format = NULL,
  serverURL = NULL,
  version = NULL,
  meta_dbattributes = NULL
)
}
\arguments{
\item{studyId}{study or experiment ID}

\item{format}{type of data structure format}

\item{serverURL}{database server URL}

\item{version}{api version}

\item{meta_dbattributes}{data dictionary of metadata. It includes equivalences between excel and database names.}
}
\description{
get metadata from personnel metadata
}
\examples{
\dontrun{
b2 <- get_personnel_metadata(studyId  = 3,format = "data.frame",
serverURL = "https://research.cip.cgiar.org/agrofims/api/dev",version ="/0233/r") 
}
}
