% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsite_metadata.R
\name{get_sitedesc_metadata}
\alias{get_sitedesc_metadata}
\title{Get metadata from experiment site information}
\usage{
get_sitedesc_metadata(
  expsiteId = NULL,
  format = NULL,
  serverURL = NULL,
  version = NULL,
  meta_dbattributes = NULL
)
}
\arguments{
\item{expsiteId}{or experimente site ID}

\item{format}{type of data structure format}

\item{serverURL}{database server URL}

\item{version}{api version}

\item{meta_dbattributes}{data dictionary of metadata. It includes equivalences between excel and database names.}
}
\description{
get metadata from experimental details
}
