% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{has_agronomic_metadata}
\alias{has_agronomic_metadata}
\title{Internal function to detect if data}
\usage{
has_agronomic_metadata(.data)
}
\arguments{
\item{.data}{data retrieved from AgAPI}
}
\description{
check anomalies in agronomic data after retrieving information from AGROFIMS database
}
\details{
Minimal conditions are 1 row, 2 columns, not all missing data
}
