% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_levels_complete}
\alias{is_levels_complete}
\title{Detect if levels are missing}
\usage{
is_levels_complete(.data)
}
\arguments{
\item{.data}{levels of experimental design factors}
}
\description{
Detect if levels are missing
}
