% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{mutate_categorical_validation_fbapp}
\alias{mutate_categorical_validation_fbapp}
\title{Mutate variable validation for categorical variables for KDSmart mobile application}
\usage{
mutate_categorical_validation_fbapp(traitlist)
}
\arguments{
\item{traitlist}{data.frame trait list data}
}
\description{
mutate validation rules according different types of inputs in order to complaint with Field Book App.
}
