% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fertilizer_measurements.R
\name{mutate_fertprod_nutelement}
\alias{mutate_fertprod_nutelement}
\title{Mutate nutrient element list}
\usage{
mutate_fertprod_nutelement(fertilizer)
}
\arguments{
\item{fertilizer}{data.frame table of fertilizer}
}
\description{
AgroFIMS retrieve in one column all the nutrient elements concatenated by pipes \code{|}. This function splits in multiple
columns where each columns is a nutrient element respectively.
}
