% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{mutate_format_fbapp}
\alias{mutate_format_fbapp}
\title{Mutate \code{format} data type attribute for Field Book App compliant}
\usage{
mutate_format_fbapp(traitlist)
}
\arguments{
\item{traitlist}{data.frame trait list table with all the variables to be measured in the field}
}
\description{
Mutate \code{format} data type attribute for Field Book App compliant
}
