% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{mutate_level_unit}
\alias{mutate_level_unit}
\title{Mutate levels and units}
\usage{
mutate_level_unit(.data)
}
\arguments{
\item{.data}{experimental design data}
}
\description{
when there exist factors which unis of measurements, this function concatenate them in one level. Ex. 30kg.
}
