% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{mutate_nummeasurement_phenology}
\alias{mutate_nummeasurement_phenology}
\title{Mutate number of measurement per season and per plot}
\usage{
mutate_nummeasurement_phenology(traitlist)
}
\arguments{
\item{traitlist}{trait list table}
}
\description{
set atributes samplesperseason and samplesperplot to 1 evaluations by default
}
