% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrient_measurements.R
\name{mutate_nutrient_split}
\alias{mutate_nutrient_split}
\title{Mutate nutrient split element list}
\usage{
mutate_nutrient_split(nutrient)
}
\arguments{
\item{nutrient}{data.frame table of nutrient}
}
\description{
AgroFIMS retrieve in one column all the nutrient split elements concatenated by pipes \code{|}. This function splits in multiple
columns where each columns is a nutrient split element respectively.
}
