% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{mutate_timming_values}
\alias{mutate_timming_values}
\title{Mutate timiing values in trait list tables}
\usage{
mutate_timming_values(traitlist)
}
\arguments{
\item{traitlist}{data.frame trait list data}
}
\description{
Traitlist tables includes timing values such as \code{Days after plating}, \code{Growth stage}, \code{Date} and \code{Frequency}. Those values are dispersing in multiple columns, so 
the idea is place this value in the \code{meaTimeValue} attribute that store timing values from different type of timing inputs.
}
