% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{mutate_variable_type_odk}
\alias{mutate_variable_type_odk}
\title{Mutate variable data type attribute for ODK mobile application}
\usage{
mutate_variable_type_odk(traitlist)
}
\arguments{
\item{traitlist}{data.frame trait list data}
}
\description{
Mutate and assign in variableDataType attribute all the ODK validations according to the type of variable (DECIMAL, INTEGER, NUMERIC,DATE,TEXT)
}
