% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{mutate_variable_validation_kdsmart}
\alias{mutate_variable_validation_kdsmart}
\title{Mutate variable validation attribute for KDSmart mobile application}
\usage{
mutate_variable_validation_kdsmart(traitlist)
}
\description{
Mutate and assign in variableValidation attribute all the kdsmart validations according to the type of variable (DECIMAL, INTEGER, NUMERIC,DATE,TEXT)
}
