% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{odk_settings_structure}
\alias{odk_settings_structure}
\title{Create ODK settings structure}
\usage{
odk_settings_structure(expsiteId = "DS92390CMNHGTO")
}
\arguments{
\item{expsiteid}{character experiment site Id}
}
\description{
Create ODK settings structure
}
