% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{odk_survey_structure}
\alias{odk_survey_structure}
\title{Create ODK survey structures for only one row rules}
\usage{
odk_survey_structure(
  type = "",
  name = "",
  label = "",
  hint = "",
  read_only = "",
  constraint = "",
  appearance = ""
)
}
\arguments{
\item{type}{character ODK type attribute}

\item{name}{character ODK name attribute}

\item{label}{character ODK label attribute}

\item{hint}{character ODK hint attribute}

\item{read_only}{character ODK read_only attribute}

\item{constraint}{character ODK constraint attribute}

\item{appeareance}{character ODK constraint appeareance}
}
\description{
Create ODK survey structures for only one row rules
}
