% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_traitlist.R
\name{remove_duplicates_traitlist}
\alias{remove_duplicates_traitlist}
\title{Remove duplicates in trait list tables}
\usage{
remove_duplicates_traitlist(traitlist)
}
\arguments{
\item{traitlist}{data.frame trait list table}
}
\description{
remove duplicates variables or measurements in the trait list table. This trait list is retrieve from AGROFIMS database via AgrAPI.
}
