% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{replaceNaCharacter}
\alias{replaceNaCharacter}
\title{Replace NA character in data frames}
\usage{
replaceNaCharacter(.data)
}
\arguments{
\item{.data}{data.frame to replace NA character by double quotes \code{("")}}
}
\description{
Replace NA character in data frames
}
