% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/protocol_data.R
\name{seq_protocol}
\alias{seq_protocol}
\title{Vectorized creation of sequence of number of evaluation based on initial values.}
\usage{
seq_protocol(.x)
}
\arguments{
\item{.x}{}
}
\description{
Vectorized creation of sequence of number of evaluation based on initial values.
}
