% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{set_protocol_names_kdsmart}
\alias{set_protocol_names_kdsmart}
\title{Set names for kdsmart protocol format}
\usage{
set_protocol_names_kdsmart(protocol_dt, dictionary)
}
\arguments{
\item{protocol_dt}{protocol data}

\item{dictionary}{dictionary to transform AgroFIMS attributes headers to KDSmart format}
}
\description{
Set names for kdsmart protocol format
}
