% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{agro_to_odk}
\alias{agro_to_odk}
\title{Transform from AgroFIMS to ODK format}
\usage{
agro_to_odk(traitlist, protocol)
}
\arguments{
\item{traitlist}{data.frame trait list table}

\item{protocol}{data.frame protocol table}
}
\description{
Transform from AgroFIMS to ODK format
}
