% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{assign_expunit}
\alias{assign_expunit}
\title{Assign  experimental unit}
\usage{
assign_expunit(
  .info_data,
  dbattr_mea = "length_p",
  dbattr_expunit = "length_p_unit"
)
}
\arguments{
\item{.info_data}{data.frame information}

\item{dbattr_mea}{character internal code for input dimension values. Example \code{length_p} is for plot length}

\item{dbattr_expunit}{character internal code for input units dimension. Example \code{length_p_unit} is for plot length units}
}
\description{
Assign and paste dimension values and dimension units. Example: \code{15 kg} , \code{20 ton/hec}.
}
