% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expdesign_data.R
\name{get_dsginfo_data}
\alias{get_dsginfo_data}
\title{Get metadata from experimental desing information}
\usage{
get_dsginfo_data(
  expsiteId = NULL,
  format = NULL,
  serverURL = NULL,
  version = NULL,
  meta_dbattributes = NULL
)
}
\arguments{
\item{expsiteId}{expsiteDbId or experimente site ID}

\item{format}{type of data structure format}

\item{serverURL}{database server URL}

\item{version}{api version}

\item{meta_dbattributes}{data dictionary of metadata. It includes equivalences between excel and database names.}
}
\description{
get metadata from experimental details
}
