% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{mutate_variable_name}
\alias{mutate_variable_name}
\title{Mutate multiple timing values for trait list table}
\usage{
mutate_variable_name(traitlist)
}
\description{
AgroFIMS timing values are divided in 5 attributes or columns: Days after planting, Growth stage, Frequency,
Date and Other. Therefore, this function assign in the column \code{meaTimeValue} all the different values according to each case or attribute.
}
