% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{mutate_variable_validation_fbapp}
\alias{mutate_variable_validation_fbapp}
\title{Mutate variable validation attribute for KDSmart mobile application}
\usage{
mutate_variable_validation_fbapp(traitlist)
}
\arguments{
\item{traitlist}{data.frame trait list data}
}
\description{
mutate validation rules according different types of inputs in order to complaint with Field Book App.
}
