% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobile_formats.R
\name{odk_settings_structure}
\alias{odk_settings_structure}
\title{Create ODK settings structure}
\usage{
odk_settings_structure(filename = "XXX-1234-4852")
}
\arguments{
\item{filename}{AgroFIMS file name}
}
\description{
Create ODK settings structure
}
